pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "videodigest"
version = "0.1.0"
description = "YouTube video understanding and summarization tool"
requires-python = ">=3.9"
dependencies = [
    "yt-dlp>=2024.1.0",
    "anthropic>=0.40.0",
    "openai-whisper>=20231117",
    "imagehash>=4.3.1",
    "Pillow>=10.0.0",
    "rich>=13.0.0",
]

[project.scripts]
videodigest = "videodigest.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["videodigest"]
